05-multi_document_rag_with_metadata_filter/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pypdf>=5.0.0
sentence-transformers>=3.0.0
google-genai>=1.0.0
python-multipart>=0.0.9
pytest>=8.0.0
httpx>=0.27.0
